Tngdisp: clean up compile warnings noise [display]
Make the world quiet!

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-1938
Reviewed-on: https://android.intel.com:443/260438

+++ display/tng/drv/psb_intel_display2.c
@@ -193,7 +193,7 @@ static int mdfld_intel_crtc_cursor_set(struct drm_crtc *crtc,
 	/*insert this bo into gtt */
 //        DRM_INFO("%s: map meminfo for hw cursor. handle %x, pipe = %d \n", __FUNCTION__, handle, pipe);
 
-	ret = psb_gtt_map_meminfo(dev, (void *)handle, 0, &page_offset);
+	ret = psb_gtt_map_meminfo(dev, (void *)((uint64_t)handle), 0, &page_offset);
 	if (ret) {
 		DRM_ERROR("Can not map meminfo to GTT. handle 0x%x\n", handle);
 		return ret;
